[omaha-client-integration-tests] use isolated system updater
This modifies the omaha-client-integration-tests so that it spins up a
system-updater using the the isolated cmx, rather than the one on
target (non-isolated).

This useful because otherwise, we'd have to OTA before running the test
with the most recent system-updater.

SECURITY: We need to modify the allowlists because we now have a new
system-updater cmx. This has the same security considerations as the
other integration tests that are on the allowlist, so it should be ok.

Test: fx test -o omaha-client-integration-tests
Bug: 64602
Change-Id: I9811373009d931fd2322981b35fb56f137c8d0d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458901
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Testability-Review: Ben Keller <galbanum@google.com>