[omaha-client] allow rebooting to clear NO_SPACE
The final CL in the sequence for fxbug.dev/65492.

If the observer has set previous_out_of_space_failure in the
FidlServer and we get a call to `PerformPendingReboot`, we should
immediately reboot. This allows us to clear a condition in which a
previous update resolved a bunch of packages and ran out of space, but
we can't clear the packages with a garbage collection because the
packages are still in pkgfs' dynamic index.

This solution exists as a short term fix. The "more correct" approach
is coming when pkg-cache has a concept of retained packages (see
fxbug.dev/60169).

Fixed: 65498
Test: add a unit test for PerformPendingReboot, fix previous unit
tests for that function
Test: add an integration test for the whole flow from pkg-resolver's
NO_SPACE to rebooting

Change-Id: Ie1f09cd1ba77645b3138eda8c2fef0669dab3b8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458777
Commit-Queue: John Wittrock <wittrock@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Zach Kirschenbaum <zkbaum@google.com>
Testability-Review: Zach Kirschenbaum <zkbaum@google.com>