[SetUI][MessageHub] Roles Introduction
Roles are a way to group messengers together under
a label (or role signature) so that messages can be
targeted towards the group as a whole. This provides
a way to target messengers in a way that is broader
than addresses and more specific than a braodcast.

Roles may be specified at compile time as part of the
MessageHub definition and/or at run time with
generated, anonymous roles.

Fixed: 64977
Test: fx test -o setui_service_tests
Multiply: setui_service_tests
Change-Id: I80f0bb11dbdb2c3b1c0d7962768cbf0ab3a56fd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453884
Commit-Queue: Bryce Lee <brycelee@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Testability-Review: William Xiao <wxyz@google.com>