[ffx][daemon] daemon.quit() safety check.
This adds a safety check to make sure not to run exit(0) inside unit
tests. This also removes unit tests that are covered in self-test which
invoke Quit as this can cause tests to exit prematurely and appear to
succeed even in the event of a failure.

Change-Id: I66523ec3756deb201ce0f0961f172513a90bdbd7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459201
Reviewed-by: Jordon Wing <jwing@google.com>
Testability-Review: Jordon Wing <jwing@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>