[storage] Wire up the FVM partition limit to fshost.
fshost currently has some configurable partition limits for blobfs and
minfs (not currently set so this code path is still unused). And FVM
supports setting limits on partition sizes.

This patch wires up the values in fshost to FVM when a device is
detected.

Separates out the mock block devices from the block-watcher-test to a
separate header so these can be shared with the new block-watcher-tests.
There is no change to these with the exception of adding a field for
the max size set.

Bug: 64055

Change-Id: I83ec188dd771e9e88ffa8b284798449640a4383b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452517
Testability-Review: Chris Suter <csuter@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Chris Suter <csuter@google.com>