[kernel][arm64][timer] add a unit test for the arm event stream timer
Validate that the event stream seems to be generating events at the rate
the cpu says it will. Err on the side of a false negative, since
anything in the system can generate events outside of the event stream,
even across cpus. Only look to see that the time spent in WFE is not too
long. Only enabled if the feature is turned.

NOTE: Running on qemu, kvm or otherwise, seems to be meaningless since
WFE either vmexits or falls through. The test will not fail, but the
measured values are extremely low, as if the instruction was largely a
nop. Potentially the same thing if running under any other hypervisor.
It may be worth simply disabling the event stream feature if the kernel
has detected its on a hypervisor, but there's not (currently) a good way
to universally detect this.

Bug: 56818
Tested: run on real hardware, validate that the observed times are
reasonable and within the computed limits.

Change-Id: I2b4a472e6848fbec760763cc9057b853870c614e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458322
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>