[brcmfmac] Add external SAE firmware feature
This change added the SAE feature flag, and reported the feature to SME
based on the flag in brcmf_if_query() and
WlanInterface::GetSupportedDrvierFeatures().
For now, in the current version firmware which supports WPA3 external
supplicant SAE, the 'extsae' feature will not show up in the feature
string requested by driver. Discussed with broadcom and they will enable
it in the next release version of the firmware.

Bug: 52234
Test: Tests added in ifc_query_test.cc and phy_query_test.cc
Change-Id: Id588c515273cef5aeba6e4d263578398a2c86728
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449821
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>
Testability-Review: Joshua Conner <joshconner@google.com>
Reviewed-by: Joshua Conner <joshconner@google.com>