[fbl] Fix FakeCallable to compile with C++20
If you try to compile the tree with -std=c++20, one of the first hurdles
you run into is fbl refusing to compile wherever fbl::Function is used
because it is only used for sizeof and doesn't have a call operator.

However, after consulting some C++ experts, it sounds like instantiating
the call operator is perfectly legal for the compiler to ask for based
on how things are now, and it's just that C++20 mode happens to be a bit
stricter (i.e. according to the standard this was always wrong).

Compiler errors are never fun, so let's nip this in the bud since it's
used so many places. Unfortunately it can't be constexpr because of some
stuff in zxtest, but since it's just a declaration you should get a link
error if anyone actually tries to use it.

TEST: no behavior change
Change-Id: I2281c913c2d4ff90e0effbdbc80fe66f919754f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459940
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>
Commit-Queue: Mark Schott <schottm@google.com>