[logs] Refactor the structured log encoder into its own header file
Expose the structured log encoder externally by creating a
header-only library for the encoder that is part of the
structured logging backend. This will eventually be used
in macros.h to replace the encoding that currently
takes place in the backend.

Change-Id: Ie18238a0afc5a8035847da0656a098571b715fa1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440534
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Testability-Review: Adam Perry <adamperry@google.com>