[time][docs] Add initial documentation structure for time.
This creates the full structure but content on many of the child pages
is not yet populated in the interest of keeping small CLs and easy
review. The remaining content will be populated by Satsuki and I over
the next 1-2 weeks.

For convenience, rendered HTML is available at
https://fuchsia.devsite.corp.google.com/preview/jsankey/

Bug: 65727
Change-Id: Iba1d3b6a46d63174eff5e56438cf50ae89149a31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458658
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Meggin Kearney <mkearney@google.com>
Reviewed-by: John Grossman <johngro@google.com>