[build][asan] Use -asan-use-private-alias=1
A change in how dependencies are propagated by GN through Rust libraries
causes:

[343.591] 284345.284347> ERROR: AddressSanitizer: odr-violation (0x23546df33ce0):
 [343.633] 284345.284347>   [1] size=29 '{{{symbol:_ZTSN3fsl13SocketDrainer6ClientE}}}' ../../src/lib/fsl/socket/socket_drainer.cc
 [343.633] 284345.284347>   [2] size=29 '{{{symbol:_ZTSN3fsl13SocketDrainer6ClientE}}}' ../../src/lib/fsl/socket/socket_drainer.cc

which has been causing the GN roll to fail for a while now.

That symbol is internal to the compiler, demangles to "typeinfo name for
fsl::SocketDrainer::Client".

Making typeinfo private to the module seems reasonable (?), which is I
believe what this clang flag does (and what gcc does by default). It
appears this is the plan for clang to default to in
https://reviews.llvm.org/D92078, and this appears to resolve the ASAN
ODR violation in local testing.

Bug: 66129
Change-Id: Iba65a2a77cdf9bde9179028ba9f9a353efb25c02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460659
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Scott Graham <scottmg@google.com>