[stdcompat] Add version header.
Provide a <version> polifyll with 'string_view' feature test symbol,
so the header is introduced in a different cl.

The goal is whenever a polyfill is added, the set of feature test macros
that are needed, are either provided by the platform or defined in this
header.

Added test that verifies for std14,17 and 20 that the symbol is either
undefined(cpp14), or at a specific draft version(17, 20) depending on
the platform cplusplus version.

Bug: 57356 #comment version polyfill

TEST=stdcompat-test(package), stdcompat-unittest(host)

Change-Id: I9834c4faabd94e0665f2de468cefa4a12139ffc9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458957
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>