[settings] Fix volume earcons via voice
Because of the async code and the recent change from timestamps to
flags, the multiple calls per-stream via voice would flip the media flag
back to the same value, and not register a change. This changes the
flags to a counter.

It also renames the corresponding flag references to "counter"

Fixed: b/175038112
Test: fx test -o setui_service_tests

Change-Id: I373b7d6dc7e2e753ee4c04da2ff182ef84490a33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460601
Reviewed-by: William Xiao <wxyz@google.com>
Testability-Review: William Xiao <wxyz@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>