[netstack] Rewrite DHCP to avoid UDP sockets
Use a packet socket to receive all traffic. This paves the way for
turning off the broadcast flag on outgoing messages, since we are now
able to receive IP packets for which we don't yet hold the address.

Write outbound traffic directly through the stack, rather than using a
UDP socket. This avoids the need to assign the unspecified address,
which causes calls to fuchsia.net.routes/State.Resolve with the
unspecified address to resolve to the local interface on which DHCP is
running. Add a test for this case to netstack_dhcp_integration_test.

Bug: 66094
Change-Id: I0416bcfca576494b3a3f627c72ddcf4136ff53a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450114
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Tamir Duberstein <tamird@google.com>
Testability-Review: Ghanan Gowripalan <ghanan@google.com>