[kernel] kcounters: Convert ifdef'ed atomic to atomic_ref
kcounters provides a way for the kernel to export statistics to
userspace, via a well-known VMO with a well-known format. The
implementation internally is a series of uint64_t counters and
descriptors, which can be aggregated to get a count of events.

The counters were updated atomically under certain conditions (relaxed
atomic add on arm64) but not in others (plain += on x86); convert them
to a lossy-add sequence -- an atomic load, add, atomic store (relaxed);
this eliminates a C++ data race and is iso-performance to the existing
code. The lossy-sequence uses atomic_ref, to not affect the
representation of the field in the exported VMO.

This change also allows removing two more legacy atomics from
kernel/atomics.h, towards fixing bug 47117.

(This language-level data race was detected by KCSAN).

Bug: 47117 Use only ktl::atomic in the kernel
Bug: 61451 Kernelmode Concurrency / Thread Sanitizer for Zircon

Change-Id: I479be11ff912ce0f96334315defc3922f56f7c00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459854
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Testability-Review: George Kulakowski <kulakowski@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>