[kernel][lib/arch] Have FakeCpuidIo return zeroed data for unknown leaves
It turns out to be too strict a requirement to expect users to
front-load data for every accessed leaf (e.g., with the xsave leaves,
which need to access subleaf 63). We update the behavior to be in line
with arch::BootCpuidIo (as well as the hardware itself in many cases).

Change-Id: Ic974d78532f48e7b2fac4c61da505cf3d39e5c44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452917
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Testability-Review: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>