[graphics][compute] Use glslangValidator `--quiet` switch
The glslangValidator compiler was very noisy by default and would
print the shader's filename whether the shader compiled cleanly or
not.

The `-s` option resolved this but would squelch *all* messages so if
there was a compile error you would have to remove the `-s` switch to
find the offending file and its error line number.

In July the `--quiet` option was added.

This appears to resolve the above issues and use of `-s` can be
dropped.

Testability: I rebuilt a Vulkan shader library and didn't see any
verbose printing.

Change-Id: I7c021f2b66ab808b293ec38c017c9f0e5ddb3e2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461674
Reviewed-by: Allan MacKinnon <allanmac@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: Allan MacKinnon <allanmac@google.com>