[blobfs] Label VMOs with merkle roots
Previously, blob VMOs were named based on the inode of the blob.

This prevents collisions, which is nice, but it makes it difficult to
figure out which VMO maps to what file. Morever, it makes it *really*
hard, nigh impossible, to figure out which VMO maps to what file in
situations where the device is not accessible, such as when looking at a
crash report. (Inodes are not stable across all devices.)

Change this to use the first 8 characters of the merkle root hash, which
is a stable identifier for the given contents. This introduces the risk
of collision, but collisions do not have any effect on blobfs' correct
functioning, and downstream memory analysis tools can break ties using
other information such as koid.

This results in vmos like `blob-d25911f9` and `mrkl-d25911f9`.

Tested: Trivial change; tested locally.
Change-Id: Icb502d80c189cad3e11828bfbd95c2b6ed13646c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457889
Fuchsia-Auto-Submit: James Sullivan <jfsulliv@google.com>
Reviewed-by: David Moore <davemoore@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>