[fdio] Fix slack in timerfd
Sometimes, waits on timerfd triggered before the timerfd was actually
readable. This was because the slack was not properly set on the
underlying timer: they were intended to use ZX_TIMER_SLACK_LATE, but
this flag was passed as the "slack" parameter (normally, a duration)
of zx_timer_set. Instead, this option should be passed to
zx_timer_create.

Bug: b/175242584
Test: stopped the following from eventually failing:
  fx test fdio-unittest-package -- \
  --gtest_filter=TimerFDTest.RepeatingNonblocking --gtest_repeat=-1 \
  --gtest_break_on_failure
Change-Id: I51963d46b8ea1be36fca74f3315d489f7da7fe20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460816
Commit-Queue: Etienne J. Membrives <etiennej@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Testability-Review: Adam Barth <abarth@google.com>