[lapis] Add FTL configurations for Sampler.
This is a reland of a configuration that
was landed and rolled back due to performance regressions.

The performance regressions were caused by a bug in a client
library built by the tool, but initially the regression was
attributed to a sampling frequency set too high. The original rollback (fxr/457656)
was relanded with a significantly reduced sampling frequency, but the
performance regression was still observed in CI (missed on reland due
to only running perf tests, not longevity tests). The persistent regression
despite reduced sampling frequency led to more in-depth investigation, leading
to the underlying client library issue being resolved.

This is a reland of the original CL, with the 60s sampling rate. Perf and longevity
tests show no regression in CPU utilization.

BUG: 66080

Change-Id: I0733d802eac835f6ce7df8573a3c38c1c7407bd5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460835
Commit-Queue: Luke Nicholson <lukenicholson@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>