[inspect][rust] Refactor state towards implementing transactions
Moves the Arc<Mutex> into the State and introduces the LockedStateGaurd
which will be used by the transactions to perform a single lock() of the
State and of the VMO per transaction.

Tested: this reorganizes the code, so existing tests suffice. A new test
for the LockedStateGuard was added in place of the test for the now gone
with_header_lock macro.

Change-Id: I67ba9581a9a724d16c558a8019983fa014e49f80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459503
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Testability-Review: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>