[codec] Log stream buffer AllocationError(s)
With this change, I don't know of any video decoder creation
allocation failure paths that lack a logged cobalt event.

If the main input or output buffer collection fails to
allocate, that'll be InputBufferAllocationFailure or
OutputBufferAllocationFailure.  If an internal buffer fails
to allocate, that'll be AllocationError.  If it's some other
non-allocation-related error, that will be a different event.

BUG: b/173512120
Change-Id: I327eb4bdcbc4861a226622b562f06913382a38aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461416
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Wez . <wez@google.com>
Testability-Review: John Bauman <jbauman@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>