[stdcompat]: Branch fit::optional.
Branch optional and test. This is a copy, before we proceed into
updating so is an accurate polyfill of optional.

The file is not exposed as part of stdcompat yet, and is not built. This
will be updated in a following patch

Bug: 57349
TEST=None

Change-Id: I4432d5924d7ae0262383bbb9d2e0c21f72a37aad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461525
Reviewed-by: Corey Tabaka <eieio@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>