[stdcompat]: Branch fit::variant.
Branch variant and test. This is a copy, before we proceed into updating
so is an accurate polyfill of variant.

The file is not exposed as part of stdcompat yet, and is not built. This
will be updated in a following patch.

Bug: 57349
TEST=NONE

Change-Id: I600d529110e941a75470998148ffdd1868c2c6cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461896
Reviewed-by: Corey Tabaka <eieio@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>