[ffx] Add interface names to link local addrs.
If an address is link local, adds the interface name. If an interface is
unplugged and cannot be shown, for now just shows the index instead of
the interface name.

Hack to get tests working on bots: interface names are given very high indexes
in order to make it highly unlikely that they will be assigned names.

Change-Id: Ica55364489ec3615be176659c75728994e867411
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460846
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Testability-Review: Andrew Davies <awdavies@google.com>