[fx] Add ninja-executable to gn gen args
As of version 1.8, ninja caches mtimes in the .ninja_log file and will
not restat a file if it has an entry in the .ninja_log. This becomes
problematic for generator files such as build.ninja. In a scenario where
the .ninja_log has an entry for build.ninja (ie, after ninja has had to
trigger a re-gen) and a manual run of of `gn gen` has updated the
build.ninja file, a subsequent invocation of ninja will trigger yet
another re-gen of build.ninja.

This can be reproduced like this:

  # Initial build
  fx set core.x64 && fx build
  # Trigger ninja re-gen so there's a build.ninja log entry
  touch BUILD.gn && fx build
  # Manual re-gen
  fx set core.x64
  # The following command will re-gen again before building
  fx build

With a recent gn change[1], gn will now use `ninja -t restat
build.ninja` to force ninja to update the .ninja_log with the accurate
mtime after generation. Gn will do this by default, so long as the user
provides a --ninja-executable to use and the ninja_required_version is
>= 1.10.0 (which is when the restat tool was added). Eliminating the
redundant generation saves 10-11 seconds during the fx build.

[1] https://gn-review.googlesource.com/c/gn/+/10400

Bug: 50061

Change-Id: I7ad53134bb4d404c574e4a36a85d4be8cdf2d738
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461355
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Reviewed-by: David Turner <digit@google.com>
Testability-Review: Shai Barack <shayba@google.com>
Commit-Queue: RJ Ascani <rjascani@google.com>