[docs][fidl] State that Rust bindings do not support struct defaults
We do not plan to support default values for struct members in the FIDL
Rust bindings because there is no easy way to implement it. A method
returning a struct instance (via std::default::Default or otherwise)
would need to pick values for the fields without defaults, such as 0 or
`Handle::invalid()`. This would go against good Rust style, which seeks
to avoid these in-between invalid states.

There are alternatives, such as only supporting defaults if _all_ fields
have a default, or using a macro instead of a method. We are
choosing instead to explicitly state that defaults are not supported.

Test: docs change
Fixed: 8013
Change-Id: I7bff6a4d87f5ecbb34712f2311c5bb6fec52962a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/462676
Reviewed-by: Felix Zhu <fcz@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>