[chromium][vulkan] Convert to Vulkan cmx shard.
Add a new cmx shard that's similar to test-application.shard.cmx, but
doesn't add the vulkan feature to the sandbox. This is useful for test
components that don't use Vulkan directly, but instead launch other
components that need to use Vulkan, and so need to pass the Vulkan
services down.

This cmx shard is useful to ensure that all tests in this situation have
access to the set of services that the Vulkan driver needs to operate.

Bug: 64508
Change-Id: I6f60de3ad544bf6bc5b8a1f6df0302b6a8fdcd6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461761
Reviewed-by: Sergey Ulanov <sergeyu@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Testability-Review: John Bauman <jbauman@google.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>