[fshost] Fix logging to serial
Opens handle to log write service and reconfigures syslog to use that
handle for logging. This is a short the fix for a bug where in on a
board with userdebug build, no logs show up on serial.

Test: Monitored serial for the messages.
Bug: 66263

Change-Id: Icb6f65484a09c12dc5eecfabba6c47655be861c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461976
Commit-Queue: Vikram Auradkar <auradkar@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>