[harvester] Gather static annotations for control panel
We want information about the device for a new FDT feature. We are
already using the build-info data for the snapshot fuchsia version, this
just reads some of the other data.

This is to unblock us until the feedback service exposes a table with
the values that we want.

Change-Id: I8bb74201d2ee37f62eb4d7dd75c9b12ac0c75338
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450236
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Jayson Adams <shrike@google.com>
Reviewed-by: Alex Bogart <abbogart@google.com>
Testability-Review: Jeremy Manson <jeremymanson@google.com>