[cts] Verify Deps are in released in an SDK
This CL changes how the CTS verifies that a dependency is released in an
SDK. Previously, the CTS would look for a {dep}_sdk.meta.json file. This
file may not exist at the time of CTS verification, and it only verified
that a dependency was an SDK atom, not that it was actually released in
an SDK. This CL introduces an implicit dependency on public/partner SDKs
to the CTS. These dependencies will generate all the SDK manifests for a
particular toolchain in the out directory. GN will wait for these
manifests to be generated before running the CTS dep verification
script.

Change-Id: Icb5c86a7bfabb2fa2d81c38fbf36943f811e91ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453857
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: Victor Chang <vfcc@google.com>
Testability-Review: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: John Shamoon <johnshamoon@google.com>