[storage] Add support for oldest version to Cobalt metrics.
To make this work, I wanted to make the cobalt collector private, which
meant removing support for custom metrics which isn't currently used. We
can add support back in later if we need to.

I looked at adding a unit test for the Blobfs change but it looks to be
non-trivial. Since the change to Blobfs is trivial and we do not seem to
be testing any of the other Cobalt metrics from Blobfs I figured it
isn't worth the effort.

Bug: 64651
Change-Id: I21bdbbefbeab17b05d48d5c265cb5c2071c56c64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460911
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Testability-Review: Vikram Auradkar <auradkar@google.com>