[brcmfmac] Fix scanning race conditions
It was possible for the driver to end up in a state where it was
continually scanning and immediately signaling that the scan was aborted
because of how transaction id's were used. When starting a scan the
transaction id for the scan would be stored as the current on-going scan
even if another scan was already in progress and the new scan was
rejected. This meant that the results, or any aborts, would be signaled
for the wrong scan. The driver could also report results for the wrong
scan because of the transaction id being stored as global state. This
created a very fragile situation where anything that would update the
global state would result in scan results being reported for the wrong
scan. Thirdly there was an issue with signaling aborted scans right
after aborting them, allowing for new scans to start. If a new scan
started before firmware also sent the abort event then the firmware
abort event would signal the first scan result twice, which is fine, but
it would also abort the new scan by resetting other global state
associated with the new scan. In this situation when the results came in
for the new scan there would be no information about a scan going on and
the results would never be sent anywhere. This would cause the upper
layers to wait indefinitely for the scan results to arrive and never
recover or attempting to scan again, including join scans, preventing
reconnects.

This change only stores the transaction id if the scan actually starts,
it also adds an atomic counter for the sync_id in the scan request to
firmware. The sync_id that was used for the request is then stored along
with the transaction id. When the sync_id is sent back in the firmware
event with the scan results we can reject scan results that do not match
the currently on-going scan. On top of that this change also modifies
the behavior so that when we abort a scan, for example when starting a
soft AP, we do not immediately signal the abort to the upper layers.
Instead we wait for the firmware event to signal the correct scan. This
way a new scan cannot start until we are completely done with the
current scan and there should be no further firmware events associated
with the old scan that can break the global scan state.

BUG: b/173283408
BUG: 65069
Test: Hours of asynchronous scanning, disconnects and reconnects
      Automated test in accompanying CL
Change-Id: I3fff9d71f96ccc055ad4a5be6801b3822858598c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457035
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Joshua Conner <joshconner@google.com>
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>